+++ CHANGELOG.md
@@ -52,6 +52,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
    - anubhav will finish this!
 - more config and metrics for rate limits and caching (#113)
 - more metrics and attempts at improvements (#114)
+- old middle queue host cooldown combined with rate limit (#117)
 
 ### Fixed
 - Fix copying body data from buffer (#24)